Split up the intermediate.rs storage part for now
